git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C353

+++ WebAPI/API/C_Development/Work/Winston/Emar.Core/Options/Model/OptionDto.cs
@@ -28,7 +28,7 @@ namespace Emar.Core.Options.Model
     {
         public int Id { get; set; }
 
-        public int SiteId { get; set; }
+        public long SiteId { get; set; }
 
         public int OptionId { get; set; }
