+++ WebAPI/API/C_Development/Work/Winston/Emar.Core/Options/Service/IOptionService.cs
@@ -2,6 +2,6 @@
 {
     public interface IOptionService
     {
-        public string GetOption(int siteId, string optionName);
+        public string GetOption(long siteId, string optionName);
     }
 }
