+++ WebAPI/API/C_Development/Work/Winston/Emar.Core/Orders/Service/OrderService_MockMethods.cs
@@ -64,7 +64,7 @@ namespace Emar.Core.Orders.Service
             return setupOptions;
         }
 
-        public IEnumerable<FrequencyDto> GetFrequencies(int siteId)
+        public IEnumerable<FrequencyDto> GetFrequencies(long siteId)
         {
             return new[]
             {
@@ -87,7 +87,7 @@ namespace Emar.Core.Orders.Service
             };
         }
 
-        public IEnumerable<UnitDto> GetUnits(in int siteId)
+        public IEnumerable<UnitDto> GetUnits(in long siteId)
         {
             return new[]
             {
